Refactor FastAPI application structure and enhance database management
- Updated database connection handling in `database.py` with a new `get_db` function for session management.
- Revamped `README.md` to provide comprehensive features, setup instructions, and project structure.
- Removed obsolete files and directories related to the previous application structure.
- Adjusted user model and schemas to align with the new database design.
