+++ alembic/versions/ae1383a808b0_add_auth_columns_to_users_table.py
@@ -1,0 +1,36 @@
+"""Add auth columns to users table
+
+Revision ID: ae1383a808b0
+Revises: 
+Create Date: 2025-07-28 21:55:35.167884
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'ae1383a808b0'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint(op.f('users_email_key'), 'users', type_='unique')
+    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
+    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_users_id'), table_name='users')
+    op.drop_index(op.f('ix_users_email'), table_name='users')
+    op.create_unique_constraint(op.f('users_email_key'), 'users', ['email'])
+    # ### end Alembic commands ###
