+++ database.py
@@ -1,10 +1,15 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, declarative_base
 
-
-
 DATABASE_URL="postgresql://postgres:1234@localhost:5432/scryps_db"
 
 engine=create_engine(DATABASE_URL, echo=True)
 SessionLocal=sessionmaker(autocommit=False, autoflush=False, bind=engine)
 Base=declarative_base()
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
