--- fastapi_auth_project/app/api/v1/routers.py
@@ -1,5 +1,0 @@
-from fastapi import APIRouter
-from .endpoints import auth
-
-api_router = APIRouter()
-api_router.include_router(auth.router, prefix="/auth", tags=["auth"])
