--- fastapi_auth_project/app/core/models.py
@@ -1,11 +1,0 @@
-from sqlalchemy import Column, Integer, String, DateTime
-from app.core.database import Base
-
-class User(Base):
-    __tablename__ = "users"
-    id = Column(Integer, primary_key=True, index=True)
-    email = Column(String, unique=True, index=True)
-    hashed_password = Column(String)
-    refresh_token = Column(String, nullable=True)
-    otp = Column(String, nullable=True)
-    otp_expiry = Column(DateTime, nullable=True)
