--- fastapi_auth_project/app/core/schemas.py
@@ -1,24 +1,0 @@
-from pydantic import BaseModel, EmailStr
-
-class UserBase(BaseModel):
-    email: EmailStr
-
-class UserCreate(UserBase):
-    password: str
-
-class User(UserBase):
-    id: int
-    class Config:
-        orm_mode = True
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-    refresh_token: str
-
-class OTPCreate(BaseModel):
-    email: EmailStr
-
-class OTPLogin(BaseModel):
-    email: EmailStr
-    otp: str
