+++ routers.py
@@ -1,0 +1,5 @@
+from fastapi import APIRouter
+from endpoints.auth import router as auth_router
+
+api_router = APIRouter()
+api_router.include_router(auth_router, prefix="/auth", tags=["auth"])
